app/models/OutgoingCall.java
package models;

import play.db.jpa.Model;

import javax.persistence.Entity;

@Entity
public class OutgoingCall extends Model {
    private String cid;
    private CallStatus status;
    private String callTo;
    private String callFrom;

    public String getCid() {
        return cid;
    }

    public void setCid(String cid) {
        this.cid = cid;
    }

    public CallStatus getStatus() {
        return status;
    }

    public void setStatus(CallStatus status) {
        this.status = status;
    }

    public String getCallTo() {
        return callTo;
    }

    public void setCallTo(String callTo) {
        this.callTo = callTo;
    }

    public String getCallFrom() {
        return callFrom;
    }

    public void setCallFrom(String callFrom) {
        this.callFrom = callFrom;
    }
}
